refactor: move exceptions.py to project root
