+++ pinterest_dl/exceptions.py
